+++ Tests/GRDBTests/RowFetchTests.swift
@@ -3,13 +3,24 @@ import GRDB
 
 class RowFetchTests: GRDBTestCase {
 
+    func testStatementInformationFromCursor() throws {
+        // Test that cursor provides statement information
+        let dbQueue = try makeDatabaseQueue()
+        try dbQueue.inDatabase { db in
+            let request: SQLRequest<Row> = "SELECT \("Arthur") AS firstName, \("Martin") AS lastName"
+            let cursor = try request.fetchCursor(db)
+            XCTAssertEqual(cursor.sql, "SELECT ? AS firstName, ? AS lastName")
+            XCTAssertEqual(cursor.arguments, ["Arthur", "Martin"])
+            XCTAssertEqual(cursor.columnCount, 2)
+            XCTAssertEqual(cursor.columnNames, ["firstName", "lastName"])
+            XCTAssertEqual(cursor.databaseRegion.description, "empty")
+        }
+    }
+    
     func testFetchCursor() throws {
         let dbQueue = try makeDatabaseQueue()
         try dbQueue.inDatabase { db in
             func test(_ cursor: RowCursor) throws {
-                // Check that RowCursor gives access to the raw SQLite API
-                XCTAssertEqual(String(cString: sqlite3_column_name(cursor.statement.sqliteStatement, 0)), "firstName")
-                
                 var row = try cursor.next()!
                 XCTAssertEqual(row["firstName"] as String, "Arthur")
                 XCTAssertEqual(row["lastName"] as String, "Martin")